Added: Damage and Health system | Minor fixes

+++ src/main/java/com/tutorial/game/JavaGame/HUD.java
@@ -4,7 +4,7 @@ import java.awt.*;
 
 public class HUD {
 
-    public int bounds = 0;
+    public float bounds = 0;
     public static float HEALTH = 100;
     private float colorTransitionValue = 255;
     private int score = 0;
@@ -19,11 +19,11 @@ public class HUD {
 
     public void render(Graphics g) {
         g.setColor(Color.GRAY);
-        g.fillRect(15, 15, 200 + bounds, 32);
+        g.fillRect(15, 15, 200 + (int) bounds, 32);
         g.setColor(new Color(75, (int) colorTransitionValue, 0));
         g.fillRect(15, 15, (int) HEALTH * 2, 32);
         g.setColor(Color.WHITE);
-        g.drawRect(15, 15, 200 + bounds, 32);
+        g.drawRect(15, 15, 200 + (int) bounds, 32);
         g.drawString("Score: " + score, 15, 64);
         g.drawString("Level: " + level, 15, 80);
         g.drawString("Space for Shop", 15, 94);
